Update reducer-helper definitions
